+++ example/lib/pages/forms/form1.dart
@@ -16,7 +16,6 @@ class Form1Component {
   int selected3;
 
   List<WSelectOption<int>> options = List.generate(20, (idx) => idx + 1)
-      .map((l) => WSelectOption<int>('Item $l', l,
-          (filter) => 'Item $l'.toLowerCase().contains(filter.toLowerCase())))
+      .map((l) => WSelectOption<int>('Item $l', l))
       .toList();
 }