+++ example/lib/pages/forms/form2.dart
@@ -18,8 +18,7 @@ class _FormData {
 ])
 class Form2Component {
   List<WSelectOption<int>> options = List.generate(20, (idx) => idx + 1)
-      .map((l) => WSelectOption<int>('Item $l', l,
-          (filter) => 'Item $l'.toLowerCase().contains(filter.toLowerCase())))
+      .map((l) => WSelectOption<int>('Item $l', l))
       .toList();
 
   _FormData data = _FormData();