+++ lib/components/tables/index.dart
@@ -1,2 +1,1 @@
 export 'w_table.dart';
-export 'w_action_table.dart';